requirements.txt
alembic==1.8.1
aniso8601==7.0.0
asgiref==3.6.0
blis==0.7.8
Boruta==0.3
catalogue==1.0.0
chardet==4.0.0
click==8.1.3
colorlover==0.3.0
cufflinks==0.17.3
cymem==2.0.6
Cython==0.29.14
databricks-cli==0.17.3
Django==3.2.18
django-cors-headers==3.14.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
docker==5.0.3
Flask==2.2.2
funcy==1.17
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.27
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.3
htmlmin==0.1.12
idna==2.10
ImageHash==4.2.1
imbalanced-learn==0.7.0
itsdangerous==2.1.2
kmodes==0.12.1
lightgbm==3.3.2
llvmlite==0.37.0
missingno==0.5.1
mlflow==1.28.0
mlxtend==0.19.0
multimethod==1.8
murmurhash==1.0.8
nltk==3.7
numba==0.54.1
oauthlib==3.2.2
pandas-profiling==3.2.0
phik==0.12.2
plac==1.1.3
preshed==3.0.7
prometheus-flask-exporter==0.20.3
promise==2.3
psycopg2-binary==2.9.5
pycaret==2.3.10
pydantic==1.10.1
PyJWT==1.7.1
pyLDAvis==3.2.2
pynndescent==0.5.7
pyod==1.0.4
pytz==2022.7.1
pywin32==227
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.25.1
Rx==1.6.3
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
singledispatch==4.0.0
six==1.16.0
smart-open==6.1.0
smmap==5.0.0
spacy==2.3.7
SQLAlchemy==1.4.40
sqlparse==0.4.2
srsly==1.0.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
text-unidecode==1.3
textblob==0.17.1
thinc==7.4.5
typing_extensions==4.5.0
umap-learn==0.5.3
urllib3==1.26.15
visions==0.7.4
waitress==2.1.2
wasabi==0.10.1
websocket-client==1.4.0
Werkzeug==2.2.2
wincertstore==0.2
wordcloud==1.8.2.2
yellowbrick==1.3.post1
